+++ enterprise/__init__.py
@@ -2,6 +2,6 @@
 Your project description goes here.
 """
 
-__version__ = "3.4.27"
+__version__ = "3.4.28"
 
 default_app_config = "enterprise.apps.EnterpriseConfig"  # pylint: disable=invalid-name